Transform to Vite
