+++ package.json
@@ -1,24 +1,21 @@
 {
   "name": "sat-estado-cfdi",
-  "version": "0.1.0",
   "private": true,
+  "version": "0.0.0",
+  "type": "module",
+  "scripts": {
+    "dev": "vite",
+    "build": "vite build",
+    "preview": "vite preview"
+  },
   "dependencies": {
     "@nodecfdi/cfdi-expresiones": "^2.1.2",
     "@nodecfdi/cfdiutils-common": "^1.2.1",
     "@nodecfdi/sat-estado-cfdi-soap": "^2.0.0",
     "@xmldom/xmldom": "^0.8.0"
   },
-  "description": "Validación ante el SAT",
-  "main": "index.js",
   "devDependencies": {
-    "webpack": "^5.74.0",
-    "webpack-cli": "^4.10.0"
-  },
-  "scripts": {
-    "build": "webpack --mode=production",
-    "dev": "webpack --mode=development",
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "author": "Hiram Abraham Villaraux Ojeda",
-  "license": "MIT"
+    "rollup-plugin-copy": "^3.4.0",
+    "vite": "^3.0.0"
+  }
 }
