--- webpack.config.js
@@ -1,8 +1,0 @@
-const path = require('path')
-
-module.exports = {
-    output: {
-        path: path.resolve(__dirname, 'build')
-    },
-    devtool: 'cheap-module-source-map'
-}
